Implement axis values and labels generators
This is meant as alternative/possible future replacement for ChartAxisValue class. With generators it's not necessary anymore to create the axis values manually, instead it's possible to specify first and last model value and let the chart generate the values. This provides more flexibility and lets also the chart control the creation of axis values.

The generation of labels was also separated from the generation of axis values, since these are completely different things. The former creates strings that display the internal numbers. The later generates the numbers.

Also:
- Add convenience initialiser to ChartAxisModel to map the fixed axis values arrays to generators, this simplifies the creation of generators for fixed arrays and provides backwards compatibility for examples.
- Remove "hidden" field from label model model, this is not necessary since by simply not generating axis value / label, the label is not shown. Remove fields in axis layers and settings from dividers and guidelines layers to filter by axis value / label visibility.