Ajust scale of chart content view on frame size change
Now when the content size is changed (i.e. label size changes due to initialisation/pan/zoom) the view is also scaled accordingly. Changing only the frame was causing offsets between chart points and axis values, since the position of the chart points remained unchaged.

Because of this change now also the min scale can't be 1 but has to be calculated dynamically. Also, since the content view now can be scaled before adding chart points the first time, it's not possible anymore to use directly screen coordinates to add views to the content view. Instead model coordinates should be used, which are mapped to screen coordinates by the layer, taking into account the current scale.